Extract NativeLibResolver
