+++ project/addons/sentry/dotnet/Sentry.Godot/Interop/NativeLibResolver.cs.uid
@@ -1,0 +1,1 @@
+uid://cwxofamjhe7nf
